Remove async and loading from GPU abstraction
Implementing async in the GPU abstraction caused many headaches with the introduction of the resource manager.
The api was originally meant to handle direct storage inside the GPU layer but it is now clear that it introduces more questions than answers, also the renderer should be directly responsible for handling platforms which support direct storage rather than bloating the api.
Callback functions for generation are still needed to reorder generation execution and make texture loading more pleasant (ex: handling mipmaps levels), this may be revisited in the future though.